+++ datadog_checks_base/CHANGELOG.md
@@ -10,6 +10,12 @@
 
 * Bump the `pyodbc` version to 4.0.39 ([#16021](https://github.com/DataDog/integrations-core/pull/16021))
 
+## 34.0.1 / 2023-10-17
+
+***Fixed***:
+
+* Fix check cancellation timeout due to `DBMAsyncJob` cancellation being blocked ([#16028](https://github.com/DataDog/integrations-core/pull/16028))
+
 ## 34.0.0 / 2023-09-29
 
 ***Changed***: